Refactor Result to be much simpler and editable.
Redo diagnostics to be a hash, ordering isn't necessary.

Redo directives as a hash with reasons.

Turn "skip" into a real status making the statuses pass, fail and skip representing the
literal outcomes of the assert.  It passed, it failed or it didn't run.

Make results boolean overloaded to return if they're considered to be a pass or a fail.

Remove the whole "result type" thing.  Formatters will have to figure that out.

Change location, id and description to file, line and name.  It was silly.

For #122